Support for Miyoo Mini in experimental branch (#207)
* New platform MMIYOO for Miyoo Mini

* Flag for 'LOADING...' message

Controlled by NOLOADING flag (0 or 1) in Makefile.

* Default option not available in Miyoo Mini

In Miyoo Mini there is no optional startup script. The closest is /mnt/SDCARD/.tmp_update/updater but we can't count on it because it is used by most distributions, so we control the quit menu manually by means of the define MM_NOQUIT.

* Flag MM_NOQUIT documentation in Makefile

* Autodelete of last_state.sav on exceptions

* System setting Audio Fix for Miyoo Mini

* Screen settings for Miyoo Mini

* Max value in color settings = 20

* Closing files that remained open

Basically those of rom_preferences that if they existed in high number could reach the limit of open files in the system/process.

* Freeing SDL ressources when critical error

Allows the application to be reopened after an exception. Previously it hung when executing SDL_SetVideoMode.